Added Next Fit (Fixed Memory Blocks) in C, C++, Java & Python
